Move ModalMixin to temba.utils.views.mixins
