+++ temba/api/views.py
@@ -13,9 +13,9 @@ from django.utils.translation import gettext_lazy as _
 from temba import mailroom
 from temba.api.support import InvalidQueryError
 from temba.contacts.models import URN
-from temba.orgs.views import ModalMixin, OrgObjPermsMixin
+from temba.orgs.views import OrgObjPermsMixin
 from temba.utils.models import TembaModel
-from temba.utils.views.mixins import NonAtomicMixin
+from temba.utils.views.mixins import ModalMixin, NonAtomicMixin
 
 from .models import APIToken, BulkActionFailure
