+++ temba/channels/types/whatsapp/views.py
@@ -10,11 +10,10 @@ from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
 
 from temba.channels.views import ChannelTypeMixin
-from temba.orgs.views import ModalMixin
 from temba.orgs.views.mixins import OrgObjPermsMixin, OrgPermsMixin
 from temba.utils.fields import InputWidget
 from temba.utils.text import truncate
-from temba.utils.views.mixins import ContextMenuMixin
+from temba.utils.views.mixins import ContextMenuMixin, ModalMixin
 
 from ...models import Channel
 from ...views import ClaimViewMixin
